+++ CHANGELOG.md
@@ -5,12 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.8.3] - Unreleased
+## [0.9.0] - Unreleased
 
 ### Added
 
+- Added textual.strip.Strip primitive
+- Added textual._cache.FIFOCache
 - Added an option to clear columns in DataTable.clear() https://github.com/Textualize/textual/pull/1427
 
+### Changed
+
+- Widget.render_line now returns a Strip
+- Fix for slow updates on Windows
+  
 ## [0.8.2] - 2022-12-28
 
 ### Fixed
@@ -308,6 +315,10 @@ https://textual.textualize.io/blog/2022/11/08/version-040/#version-040
 - New handler system for messages that doesn't require inheritance
 - Improved traceback handling
 
+[0.9.0]: https://github.com/Textualize/textual/compare/v0.8.2...v0.9.0
+[0.8.2]: https://github.com/Textualize/textual/compare/v0.8.1...v0.8.2
+[0.8.1]: https://github.com/Textualize/textual/compare/v0.8.0...v0.8.1
+[0.8.0]: https://github.com/Textualize/textual/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/Textualize/textual/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/Textualize/textual/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/Textualize/textual/compare/v0.4.0...v0.5.0